refactor(lib): migrate linked-case utility to TypeScript and update imports
